+++ omni_lang/omni_lang/tests/test65.nim
@@ -1,0 +1,11 @@
+import ../../omni_lang, macros
+
+expandMacros:
+    def something:
+        return 0.5
+
+    def something2 float:
+        return 0.5
+
+    def something3 float:
+        return 0.5